+++ scripts/lnx/osiris_common.h
@@ -1,5 +1,5 @@
 /*
-* Descent 3 
+* Descent 3
 * Copyright (C) 2024 Parallax Software
 *
 * This program is free software: you can redistribute it and/or modify
@@ -931,7 +931,7 @@ typedef unsigned char ubyte;
 typedef signed char sbyte;
 typedef unsigned short ushort;
 typedef unsigned int uint;
-typedef unsigned long ulong;
+typedef unsigned int ulong;
 typedef unsigned int ddgr_color;
 
 #ifndef NULL
